Fix sign conversions in the shared libraries
Remove all 357 clang-diagnostic-sign-conversion reports in tech, sdllib, port
and winvq, mostly by fixing types at their source rather than casting:

- tech: XMP_Count_Bits and the two-argument XMP_Encode return int, bit
  indices are int, LZW characters and LZO buffers are unsigned char byte
  data, and PKey::Fast_Exponent returns uint32_t.
- sdllib: WSA and audio sizes and sample offsets are signed, the WSA header
  size macros are constexpr ints, and Get_Random_Mask takes int.
- winvq: the VQA player's long/unsigned long fields become int32_t/int64_t
  numbers and uint32_t flags with unsigned flag constants; UnVQ takes int
  dimensions. On-disk VQA/IFF headers keep their layout.

C library lengths, allocations and container sizes go through base::ToSize
and base::ToSigned. The sdllib file seek wrappers keep plain casts because
callers pass negative SEEK_CUR offsets through size_t.

+++ src/sdllib/file.cc
@@ -51,8 +51,10 @@ bool IO_Write_File(void* handle, const void* buffer, size_t count,
 
 size_t IO_Seek_File(void* handle, size_t offset, int origin) {
   auto* file = static_cast<FILE*>(handle);
-  fseek(file, offset, origin);
-  return ftell(file);
+  // Callers pass negative SEEK_CUR offsets and receive ftell's -1 error result
+  // through size_t, so both conversions deliberately preserve the bit pattern.
+  fseek(file, static_cast<int64_t>(offset), origin);
+  return static_cast<size_t>(ftell(file));
 }
 
 size_t IO_Get_File_Size(void* handle) {
@@ -65,7 +67,8 @@ size_t IO_Get_File_Size(void* handle) {
 
   fseek(file, pos, SEEK_SET);
 
-  return length;
+  // As in IO_Seek_File, ftell's -1 error result wraps through size_t.
+  return static_cast<size_t>(length);
 }
 
 bool IO_Delete_File(const char* filename) { return unlink(filename) == 0; }
