+++ src/winvq/vqa32/unvq.h
@@ -43,11 +43,11 @@
  ****************************************************************************/
 
 void UnVQ_4x2(const unsigned char* codebook, const unsigned char* pointers,
-              unsigned char* buffer, unsigned long blocksperrow,
-              unsigned long numrows, unsigned long bufwidth);
+              unsigned char* buffer, int blocksperrow, int numrows,
+              int bufwidth);
 
 void UnVQ_4x4(const unsigned char* codebook, const unsigned char* pointers,
-              unsigned char* buffer, unsigned long blocksperrow,
-              unsigned long numrows, unsigned long bufwidth);
+              unsigned char* buffer, int blocksperrow, int numrows,
+              int bufwidth);
 
 #endif  // CNC_RED_ALERT_WINVQ_VQA32_UNVQ_H_
